Split Rfc003 default from split protocol defaults
